refactor: Move empty state into a component 8b3b471
We found that it was unclear that the `.empty-state`
stack component was supposed to be replaced rather
than used as a container for user content. So in this
commit we move the empty state into it's own
component to make it more difficult for users to get this
wrong.

newrelic/open-source-office#26